pyproject.toml
[project]
name = "scalp-radar"
version = "0.1.0"
description = "Multi-strategy automated scalping tool for crypto futures"
requires-python = ">=3.12"
dependencies = [
    "ccxt>=4.0",
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "pandas>=2.2",
    "numpy>=2.0",
    "pydantic>=2.8",
    "pydantic-settings>=2.4",
    "pyyaml>=6.0",
    "aiosqlite>=0.20",
    "websockets>=13.0",
    "loguru>=0.7",
    "python-dotenv>=1.0",
    "httpx>=0.27",
    "tqdm>=4.66",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
]
